LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/request-end-to-end.https-expected.txt


FAIL Test FetchEvent.request passed to onfetch assert_equals: Constructing a Request with a Request whose mode is navigate and non-empty RequestInit must throw a TypeError. expected "TypeError" but got ""
